Changed gold implementation from Integer to Float. Added new GoldPerSecond attribute and according systems (currently still disabled because it spams the GUI with all those gold changes every frame...).

+++ Projects/Amara-Ingame-Client/src/amara/applications/ingame/client/gui/ScreenController_HUD.java
@@ -171,8 +171,8 @@ public class ScreenController_HUD extends GameScreenController{
         getElementByID(prefix + "_" + index + "_cooldown").hide();
     }
     
-    public void setGold(int gold){
-        getTextRenderer("gold").setText("" + gold);
+    public void setGold(float gold){
+        getTextRenderer("gold").setText("" + ((int) gold));
     }
     
     public void setPing(int ping){
